requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
aws-cdk.assertions==1.139.0
aws-cdk.assets==1.139.0
aws-cdk.aws-acmpca==1.139.0
aws-cdk.aws-apigateway==1.139.0
aws-cdk.aws-apigatewayv2-integrations==1.139.0
aws-cdk.aws-applicationautoscaling==1.139.0
aws-cdk.aws-appsync==1.139.0
aws-cdk.aws-autoscaling==1.139.0
aws-cdk.aws-autoscaling-common==1.139.0
aws-cdk.aws-autoscaling-hooktargets==1.139.0
aws-cdk.aws-certificatemanager==1.139.0
aws-cdk.aws-cloudformation==1.139.0
aws-cdk.aws-cloudfront==1.139.0
aws-cdk.aws-cloudwatch==1.139.0
aws-cdk.aws-codebuild==1.139.0
aws-cdk.aws-codecommit==1.139.0
aws-cdk.aws-codedeploy==1.139.0
aws-cdk.aws-codeguruprofiler==1.139.0
aws-cdk.aws-codepipeline==1.139.0
aws-cdk.aws-codepipeline-actions==1.139.0
aws-cdk.aws-codestar==1.139.0
aws-cdk.aws-codestarnotifications==1.139.0
aws-cdk.aws-cognito==1.139.0
aws-cdk.aws-dynamodb==1.139.0
aws-cdk.aws-ec2==1.139.0
aws-cdk.aws-ecr==1.139.0
aws-cdk.aws-ecr-assets==1.139.0
aws-cdk.aws-ecs==1.139.0
aws-cdk.aws-efs==1.139.0
aws-cdk.aws-elasticloadbalancing==1.139.0
aws-cdk.aws-elasticloadbalancingv2==1.139.0
aws-cdk.aws-elasticsearch==1.139.0
aws-cdk.aws-events==1.139.0
aws-cdk.aws-events-targets==1.139.0
aws-cdk.aws-globalaccelerator==1.139.0
aws-cdk.aws-iam==1.139.0
aws-cdk.aws-kinesis==1.139.0
aws-cdk.aws-kinesisfirehose==1.139.0
aws-cdk.aws-kms==1.139.0
aws-cdk.aws-lambda==1.139.0
aws-cdk.aws-lambda-python==1.139.0
aws-cdk.aws-logs==1.139.0
aws-cdk.aws-rds==1.139.0
aws-cdk.aws-route53==1.139.0
aws-cdk.aws-route53-targets==1.139.0
aws-cdk.aws-s3==1.139.0
aws-cdk.aws-s3-assets==1.139.0
aws-cdk.aws-s3-deployment==1.139.0
aws-cdk.aws-sam==1.139.0
aws-cdk.aws-secretsmanager==1.139.0
aws-cdk.aws-servicediscovery==1.139.0
aws-cdk.aws-signer==1.139.0
aws-cdk.aws-sns==1.139.0
aws-cdk.aws-sns-subscriptions==1.139.0
aws-cdk.aws-sqs==1.139.0
aws-cdk.aws-ssm==1.139.0
aws-cdk.aws-stepfunctions==1.139.0
aws-cdk.cloud-assembly-schema==1.139.0
aws-cdk.core==1.139.0
aws-cdk.custom-resources==1.139.0
aws-cdk.cx-api==1.139.0
aws-cdk.lambda-layer-awscli==1.139.0
aws-cdk.pipelines==1.139.0
aws-cdk.region-info==1.139.0
aws_lambda_powertools
aws-wsgi==0.2.7
boto3==1.20.50
botocore==1.23.50
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
constructs==3.3.211
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsii==1.52.1
MarkupSafe==2.0.1
numpy==1.22.2
packaging==21.3
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
flask_swagger_ui